+++ common/synonyms.py
@@ -28,8 +28,13 @@ class SynonymsCollection(object):
 
     def add_synonym_value(self, value):
         assert(isinstance(value, unicode))
-        assert(not self.contains_synonym_value(value))
-        assert(not self.__is_read_only)
+
+        if self.contains_synonym_value(value):
+            raise Exception((u"Collection already contains synonyms '{}'".format(value)).encode('utf-8'))
+
+        if self.__is_read_only:
+            raise Exception((u"Failed to add '{}'. Synonym collection is read only!".format(value)).encode('utf-8'))
+
         id = self.create_synonym_id(self.__stemmer, value)
         self.__by_synonym[id] = self.__get_groups_count()
         self.__by_index.append([value])